+++ pcs/src/univariate_kzg/srs.rs
@@ -162,12 +162,16 @@ impl<E: Pairing> StructuredReferenceString for UnivariateUniversalParams<E> {
 mod tests {
     use super::UnivariateUniversalParams;
     use crate::PCSError;
-    use ark_ec::{pairing::Pairing, scalar_mul::fixed_base::FixedBase, CurveGroup};
+    use ark_ec::{
+        pairing::Pairing, scalar_mul::BatchMulPreprocessing, AffineRepr, CurveGroup, ScalarMul,
+    };
     use ark_ff::PrimeField;
     use ark_std::{
         end_timer,
         rand::{CryptoRng, RngCore},
-        start_timer, vec, One, UniformRand,
+        start_timer, vec,
+        vec::Vec,
+        One, UniformRand,
     };
 
     pub(crate) fn gen_srs_for_testing<E: Pairing, R: RngCore + CryptoRng>(
@@ -193,14 +197,12 @@ mod tests {
             cur *= &beta;
         }
 
-        let window_size = FixedBase::get_mul_window_size(prover_degree + 1);
-
-        let scalar_bits = E::ScalarField::MODULUS_BIT_SIZE as usize;
         let g_time = start_timer!(|| "Generating powers of G");
-        // TODO: parallelization
-        let g_table = FixedBase::get_window_table(scalar_bits, window_size, g);
-        let powers_of_g =
-            FixedBase::msm::<E::G1>(scalar_bits, window_size, &g_table, &powers_of_beta);
+        let powers_of_g = g
+            .batch_mul(&powers_of_beta)
+            .into_iter()
+            .map(|p| p.into_group())
+            .collect::<Vec<_>>();
         end_timer!(g_time);
 
         let powers_of_g = E::G1::normalize_batch(&powers_of_g);