+++ relation/src/gadgets/ecc/emulated/twisted_edwards.rs
@@ -165,7 +165,7 @@ mod tests {
     use ark_bn254::Fr as Fr254;
     use ark_ec::{
         short_weierstrass::{Projective, SWCurveConfig},
-        CurveGroup, Group,
+        CurveGroup, PrimeGroup,
     };
     use ark_ff::{MontFp, PrimeField};
     use ark_std::{UniformRand, Zero};