+++ relation/src/gadgets/ecc/mod.rs
@@ -598,7 +598,7 @@ mod test {
     use super::*;
     use crate::gadgets::test_utils::test_variable_independence_for_circuit;
     use ark_bls12_377::{g1::Config as Param761, Fq as Fq377};
-    use ark_ec::{twisted_edwards::TECurveConfig as Config, Group};
+    use ark_ec::{twisted_edwards::TECurveConfig as Config, AdditiveGroup};
     use ark_ed_on_bls12_377::{EdwardsConfig as Param377, Fq as FqEd377};
     use ark_ed_on_bls12_381::{EdwardsConfig as Param381, Fq as FqEd381};
     use ark_ed_on_bls12_381_bandersnatch::{EdwardsConfig as Param381b, Fq as FqEd381b};
@@ -618,7 +618,7 @@ mod test {
 
     fn test_is_neutral_helper<F, P>() -> Result<(), CircuitError>
     where
-        F: PrimeField + ?Sized,
+        F: PrimeField,
         P: Config<BaseField = F>,
     {
         let mut circuit: PlonkCircuit<F> = PlonkCircuit::new_turbo_plonk();