+++ prf/CHANGELOG.md
@@ -5,9 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.2.0
 
-- Upgrade arkworks dependencies to v0.5.0.
+- [#827](https://github.com/EspressoSystems/jellyfish/pull/827) Upgrade arkworks dependencies to v0.5.0.
 
 ## 0.1.0
 
 - Initial release.
-- AEAD wrapper for `crypto_kx`.
+- Trait definition for Pseudo-random Functions(PRF).