+++ signature/CHANGELOG.md
@@ -3,6 +3,26 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Untracked
+
+### Changed
+
+- [#842](https://github.com/EspressoSystems/jellyfish/pull/842): `jf-rescue` is now optional for the crate.
+
+## 0.4.0
+
+### Breaking Changes
+
+- [#840](https://github.com/EspressoSystems/jellyfish/pull/840): `jf-rescue` moved away from `spongefish` migration.
+
+## 0.3.0
+
+### Changed
+
+- [#827](https://github.com/EspressoSystems/jellyfish/pull/827): Upgrade arkworks dependencies to v0.5.0.
+- [#773](https://github.com/EspressoSystems/jellyfish/pull/773): Add subgroup and on curve check to BLS signatures and public keys
+- [#786](https://github.com/EspressoSystems/jellyfish/pull/786): Debug public key types as TaggedBase64
+
 ## [0.2.0]((https://github.com/EspressoSystems/jellyfish/compare/jf-signatures-v0.1.1...jf-signatures-v0.2.0)) (2024-10-29)
 
 ### Changed
@@ -15,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- [#586](https://github.com/EspressoSystems/jellyfish/pull/586) Omit private keys from logging and debug output.
+- [#586](https://github.com/EspressoSystems/jellyfish/pull/586): Omit private keys from logging and debug output.
 
 ## 0.1.0